# public/openapi.yaml
openapi: 3.0.3
info:
  title: GDP - Plateforme d'Apprentissage P2P
  description: |
    API pour la plateforme GDP permettant aux utilisateurs d'apprendre et partager des connaissances.
    Authentification via Supabase avec magic links.
  version: 1.0.0
  contact:
    name: GDP Support
    email: support@gdp.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://studxus.vercel.app/api
    description: Serveur de production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT fourni par Supabase
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sender_id:
          type: string
          format: uuid
        receiver_id:
          type: string
          format: uuid
        content:
          type: string
          description: Contenu du message
        read:
          type: boolean
          description: Message lu ou non
        sent_at:
          type: string
          format: date-time
      required:
        - sender_id
        - receiver_id
        - content
...
paths:
...
  /messages:
    get:
      tags:
        - Messages
      summary: Liste des messages
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Liste des messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      tags:
        - Messages
      summary: Envoyer un message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                receiver_id:
                  type: string
                  format: uuid
                content:
                  type: string
              required:
                - receiver_id
                - content
      responses:
        "201":
          description: Message envoyé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...